Fix : "Better code "
