--- database/db.config.js
@@ -1,10 +1,0 @@
-const pool = {
-    user: 'postgres',
-    host: 'localhost',
-    database: 'likeme',
-    password: 'sa',
-    port: 5432,
-    allowExitOnIdle: true
-}
-
-module.exports = { pool };
