+++ routes/database.js
@@ -1,12 +1,11 @@
 const { Router } = require('express')
 
-const dbController = require("../controllers/db.controller.js")
+const dbController = require("../controllers/database.js")
 
 const router = Router()
 
 router
     .get("/posts", dbController.getAll)
     .post("/posts", dbController.createNew)
-
 
 module.exports = router;
