+++ services/database.js
@@ -1,0 +1,13 @@
+const { Pool } = require('pg');
+
+const pool = new Pool({
+    user: 'postgres',
+    host: 'localhost',
+    database: 'likeme',
+    password: 'sa',
+    port: 5432,
+});
+
+const query = (text, params) => pool.query(text, params);
+
+module.exports = { query };
